DEX-339 Cache classes relocated
