+++ dex/src/main/scala/com/wavesplatform/dex/cache/AssetDecimalsCache.scala
@@ -1,4 +1,4 @@
-package com.wavesplatform.dex.model
+package com.wavesplatform.dex.cache
 
 import java.util.concurrent.ConcurrentHashMap
