+++ dex/src/test/scala/com/wavesplatform/dex/MatcherTestData.scala
@@ -7,6 +7,7 @@ import com.google.common.primitives.{Bytes, Ints}
 import com.typesafe.config.ConfigFactory
 import com.wavesplatform.account.KeyPair
 import com.wavesplatform.common.state.ByteStr
+import com.wavesplatform.dex.cache.RateCache
 import com.wavesplatform.dex.model.MatcherModel.Price
 import com.wavesplatform.dex.model._
 import com.wavesplatform.dex.queue.{QueueEvent, QueueEventWithMeta}
