+++ dex/src/test/scala/com/wavesplatform/dex/api/MatcherApiRouteSpec.scala
@@ -10,6 +10,7 @@ import com.typesafe.config.ConfigFactory
 import com.wavesplatform.account.KeyPair
 import com.wavesplatform.common.utils.Base58
 import com.wavesplatform.dex._
+import com.wavesplatform.dex.cache.RateCache
 import com.wavesplatform.dex.settings.MatcherSettings
 import com.wavesplatform.http.ApiMarshallers._
 import com.wavesplatform.http.RouteSpec
